+++ src/app/core/interface/api-response.ts
@@ -1,4 +1,3 @@
 export interface ApiResponse<T = void> {
-  message: string;
   data: T;
 }