+++ src/app/features/manage-user/page/create-page/create-page.component.ts
@@ -1,9 +1,12 @@
 import { Component } from '@angular/core';
+import { FormComponent } from '@features/manage-user/component/form/form.component';
 
 @Component({
   selector: 'app-create-page',
-  imports: [],
+  imports: [FormComponent],
   templateUrl: './create-page.component.html',
   styleUrl: './create-page.component.scss'
 })
-export class CreatePageComponent {}
+export class CreatePageComponent {
+  submit() {}
+}