+++ src/app/features/manage-user/page/update-page/update-page.component.ts
@@ -1,9 +1,12 @@
 import { Component } from '@angular/core';
+import { FormComponent } from '@features/manage-user/component/form/form.component';
 
 @Component({
   selector: 'app-update-page',
-  imports: [],
+  imports: [FormComponent],
   templateUrl: './update-page.component.html',
   styleUrl: './update-page.component.scss'
 })
-export class UpdatePageComponent {}
+export class UpdatePageComponent {
+  submit() {}
+}