+++ src/app/shared/component/select/select.component.ts
@@ -1,0 +1,23 @@
+import { CommonModule } from '@angular/common';
+import { Component, Input } from '@angular/core';
+import { FormControl, ReactiveFormsModule } from '@angular/forms';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
+
+@Component({
+  selector: 'app-select',
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatSelectModule
+  ],
+  templateUrl: './select.component.html',
+  styleUrl: './select.component.scss'
+})
+export class SelectComponent {
+  @Input() control!: FormControl;
+  @Input() label: string = '';
+  @Input() placeholder: string = 'Select option';
+  @Input() options: { label: string; value: any }[] = [];
+}