+++ src/environments/environment.ts
@@ -1,4 +1,4 @@
-const baseUrl = 'http://localhost:4200';
+const baseUrl = 'https://67d8487d00348dd3e2a6eda0.mockapi.io/user-role/api/v1';
 
 export const environment = {
   default: baseUrl